Rename test/test_utils.py into test/utils.py
